## Description:
- Updates version.py to optionally accept git commit timestamp, commit ID, and branch name from the environment. SymCrypt is built from a source tarball in Mariner and cannot get this information from from git.
- Fixed build warnings encountered in Mariner build system

## Admin Checklist:
- [x] You have updated documentation in symcrypt.h to reflect any changes in behavior
- [x] You have updated CHANGELOG.md to reflect any changes in behavior
- [x] You have updated symcryptunittest to exercise any new functionality
- [x] If you have introduced any symbols in symcrypt.h you have updated production and test dynamic export symbols (exports.ver / exports.def / symcrypt.src) and tested the updated dynamic modules with symcryptunittest
- [x] If you have introduced functionality that varies based on CPU features, you have manually tested with and without relevant features
- [x] If you have made significant changes to a particular algorithm, you have checked that performance numbers reported by symcryptunittest are in line with expectations
- [x] If you have added new algorithms/modes, you have updated the status indicator text for the associated modules if necessary

+++ unittest/exe_linux/main_exe.cpp
@@ -86,7 +86,7 @@ verifyVectorRegisters()
         // It is perfectly fine for the XMM register values to have been modified.
         // We just test that the top half of the Ymm registers have been preserved.
         //
-        for( int i=0; i<sizeof( g_ymmStartState ); i++ )
+        for( SIZE_T i=0; i<sizeof( g_ymmStartState ); i++ )
         {
             if( ((volatile BYTE * )&g_ymmStartState[0])[i] != ((volatile BYTE * )&g_ymmTestState[0])[i] &&
                 ((i & 16) == 16 )
