+++ unittest/lib/testKmac.cpp
@@ -109,7 +109,7 @@ SIZE_T KmacMultiImp::inputBlockLen()
     for(KmacImpPtrVector::const_iterator i = m_imps.begin(); i != m_imps.end(); ++i )
     {
         SIZE_T v = (*i)->inputBlockLen();
-        CHECK( res == -1 || res == v, "Inconsistent input block len" );
+        CHECK( res == (SIZE_T) -1 || res == v, "Inconsistent input block len" );
         res = v;
     }
