+++ unittest/lib/testXof.cpp
@@ -90,7 +90,7 @@ SIZE_T XofMultiImp::inputBlockLen()
     for( XofImpPtrVector::const_iterator i = m_imps.begin(); i != m_imps.end(); ++i )
     {
         SIZE_T v = (*i)->inputBlockLen();
-        CHECK( res == -1 || res == v, "Inconsistent input block len" );
+        CHECK( res == (SIZE_T) -1 || res == v, "Inconsistent input block len" );
         res = v;
     }
